+++ resolvers/record_field_resolver.go
@@ -56,7 +56,7 @@ type CollectionsFinder interface {
 //	resolver := resolvers.NewRecordFieldResolver(
 //	    app.Dao(),
 //	    myCollection,
-//	    &models.RequestData{...},
+//	    &models.RequestInfo{...},
 //	    true,
 //	)
 //	provider := search.NewProvider(resolver)
@@ -68,21 +68,21 @@ type RecordFieldResolver struct {
 	allowedFields     []string
 	loadedCollections []*models.Collection
 	joins             []*join // we cannot use a map because the insertion order is not preserved
-	requestData       *models.RequestData
-	staticRequestData map[string]any
+	requestInfo       *models.RequestInfo
+	staticRequestInfo map[string]any
 }
 
 // NewRecordFieldResolver creates and initializes a new `RecordFieldResolver`.
 func NewRecordFieldResolver(
 	dao CollectionsFinder,
 	baseCollection *models.Collection,
-	requestData *models.RequestData,
+	requestInfo *models.RequestInfo,
 	allowHiddenFields bool,
 ) *RecordFieldResolver {
 	r := &RecordFieldResolver{
 		dao:               dao,
 		baseCollection:    baseCollection,
-		requestData:       requestData,
+		requestInfo:       requestInfo,
 		allowHiddenFields: allowHiddenFields,
 		joins:             []*join{},
 		loadedCollections: []*models.Collection{baseCollection},
@@ -97,17 +97,17 @@ func NewRecordFieldResolver(
 		},
 	}
 
-	r.staticRequestData = map[string]any{}
-	if r.requestData != nil {
-		r.staticRequestData["method"] = r.requestData.Method
-		r.staticRequestData["query"] = r.requestData.Query
-		r.staticRequestData["headers"] = r.requestData.Headers
-		r.staticRequestData["data"] = r.requestData.Data
-		r.staticRequestData["auth"] = nil
-		if r.requestData.AuthRecord != nil {
-			r.requestData.AuthRecord.IgnoreEmailVisibility(true)
-			r.staticRequestData["auth"] = r.requestData.AuthRecord.PublicExport()
-			r.requestData.AuthRecord.IgnoreEmailVisibility(false)
+	r.staticRequestInfo = map[string]any{}
+	if r.requestInfo != nil {
+		r.staticRequestInfo["method"] = r.requestInfo.Method
+		r.staticRequestInfo["query"] = r.requestInfo.Query
+		r.staticRequestInfo["headers"] = r.requestInfo.Headers
+		r.staticRequestInfo["data"] = r.requestInfo.Data
+		r.staticRequestInfo["auth"] = nil
+		if r.requestInfo.AuthRecord != nil {
+			r.requestInfo.AuthRecord.IgnoreEmailVisibility(true)
+			r.staticRequestInfo["auth"] = r.requestInfo.AuthRecord.PublicExport()
+			r.requestInfo.AuthRecord.IgnoreEmailVisibility(false)
 		}
 	}
 
@@ -166,7 +166,7 @@ func (r *RecordFieldResolver) resolveStaticRequestField(path ...string) (*search
 	path[len(path)-1] = lastProp
 
 	// extract value
-	resultVal, err := extractNestedMapVal(r.staticRequestData, path...)
+	resultVal, err := extractNestedMapVal(r.staticRequestInfo, path...)
 
 	if modifier == issetModifier {
 		if err != nil {
@@ -175,7 +175,7 @@ func (r *RecordFieldResolver) resolveStaticRequestField(path ...string) (*search
 		return &search.ResolverResult{Identifier: "TRUE"}, nil
 	}
 
-	// note: we are ignoring the error because requestData is dynamic
+	// note: we are ignoring the error because requestInfo is dynamic
 	// and some of the lookup keys may not be defined for the request
 
 	switch v := resultVal.(type) {